refactor(db): rename db dir to api
