+++ manage.py
@@ -1,5 +1,5 @@
 from flask.cli import FlaskGroup
-from stockjockey import create_app, db  # , User
+from stockjockey import create_app, api  # , User
 
 app = create_app()
 cli = FlaskGroup(app)
@@ -7,9 +7,9 @@ cli = FlaskGroup(app)
 
 @cli.command("create_db")
 def create_db():
-    db.drop_all()
-    db.create_all()
-    db.session.commit()
+    api.drop_all()
+    api.create_all()
+    api.session.commit()
 
 
 # @cli.command("seed_db")
